+++ SomnusBackend/src/main/java/com/somnus/server/backend/teammembers/database/ContributionsDatabaseConnection.java
@@ -1,0 +1,13 @@
+package com.somnus.server.backend.teammembers.database;
+
+import com.somnus.server.backend.genericDatabaseConnection.ObjectDatabaseConnection;
+import com.somnus.server.backend.teammembers.domain.Contribution;
+
+public class ContributionsDatabaseConnection extends ObjectDatabaseConnection<Contribution> {
+
+    private static final String contributionsDatabaseName = "contributions";
+
+    public ContributionsDatabaseConnection() {
+        super(contributionsDatabaseName);
+    }
+}